Move the pipe-task outcome tests beside the module they test
`test_pipeline_streams_fd_lifecycle.py` had reached 425 lines, past the
400-line limit, and the tests for `_surface_unexpected_pipe_failures`
were the part that did not belong: they cover which pipe-task outcomes a
pipeline tolerates, not the raw-descriptor fault injection the rest of
the module exists for.

Move them to `test_pipeline_pipe_tasks.py`, matching the production split
between `_pipeline_pipe_tasks` and `_pipeline_stream_fds`. Both files now
sit well inside the limit at 364 and 77 lines.

Refresh the wheel manifest for the new module.
